+++ tests/browser_assert.test.js
@@ -77,6 +77,13 @@ describe("Assertions", function() {
         }, `Expected element ".container p" to have text "My second paragraph", "My first paragraph" found`);
     });
 
+    it("Multiple Text Throws", async () => {
+        await browser.open(configUrls.index);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.text("p", "My paragraph");
+        }, `Expected element "p" to have text "My paragraph", "My first paragraph My second paragraph" found`);
+    });
+
     it("Text Throws With Custom Message", async () => {
         await browser.open(configUrls.index);
         await utils.assertThrowsAsync(async () => {
@@ -280,4 +287,40 @@ describe("Assertions", function() {
         }, `element failed`);
     });
 
+    it("Text Contains", async () => {
+        await browser.open(configUrls.index);
+        await browser.assert.textContains(".container p", "My first paragraph");
+        await browser.assert.textContains(".container p", "My first");
+    });
+
+    it("Text Contains Multiple Elements", async () => {
+        await browser.open(configUrls.index);
+        await browser.assert.textContains("p", "My first paragraph");
+        await browser.assert.textContains("p", "My second paragraph");
+        await browser.assert.textContains("p", "My second");
+        await browser.assert.textContains("p", "My first");
+    });
+
+    it("Text Contains From Node", async () => {
+        await browser.open(configUrls.index);
+        const node = await browser.query(".container p");
+        await browser.assert.textContains(node, "My first");
+    });
+
+    it("Text Contains Throws", async () => {
+        await browser.open(configUrls.index);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.textContains(".container p", "My second");
+        }, `Expected element ".container p" to contain text "My second", "My first paragraph" found`);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.textContains("p", "My paragraph");
+        }, `Expected element "p" to contain text "My paragraph", "My first paragraph My second paragraph" found`);
+    });
+
+    it("Text Contains Throws With Custom Message", async () => {
+        await browser.open(configUrls.index);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.textContains(".container p", "My second", "text contains fails");
+        }, `text contains fails`);
+    });
 });